+++ tests/test_fp8.py
@@ -5,7 +5,7 @@ import sys
 import array
 import math
 import bitstring
-from bitstring import Bits, BitArray, BitStream, Dtype
+from bitstring import Bits, BitArray, Dtype, Reader
 from bitstring.fp8 import p4binary_fmt, p3binary_fmt
 from bitstring.mxfp import e4m3mxfp_saturate_fmt, e5m2mxfp_saturate_fmt, e3m2mxfp_fmt, e2m3mxfp_fmt, e2m1mxfp_fmt
 from gfloat.formats import (format_info_ocp_e4m3, format_info_ocp_e5m2, format_info_p3109, format_info_ocp_e3m2,
@@ -44,7 +44,7 @@ class TestFp8:
         assert a.p3binary == 0.0
 
     def test_reading(self):
-        a = BitStream('0x00fff')
+        a = Reader(Bits('0x00fff'))
         x = a.read('p3binary')
         assert x == 0.0
         assert a.pos == 8
@@ -55,7 +55,7 @@ class TestFp8:
     def test_read_list(self):
         v = [-6, -2, 0.125, 7, 10]
         a = bitstring.pack('5*p4binary', *v)
-        vp = a.readlist('5*p4binary')
+        vp = Reader(a).readlist('5*p4binary')
         assert v == vp
 
     def test_interpretations(self):
@@ -332,4 +332,3 @@ def test_rounding_consistent_to_gfloat():
                 assert math.isnan(theirs)
             else:
                 assert mine == theirs
-